+++ src/Fleans/Fleans.Domain/Activities/ActivityResult.cs
