+++ src/Fleans/Fleans.Domain/Activities/ActivityResultStatus.cs
