+++ src/Fleans/Fleans.Domain/Activities/GatewayExclusiveActivity.cs
