+++ src/Fleans/Fleans.Domain/Activities/IActivity.cs
