+++ src/Fleans/Fleans.Domain/Activities/IActivityExecutionResult.cs
