+++ src/Fleans/Fleans.Domain/Builders/ExclusiveGatewayBuilder.cs
@@ -1,3 +1,4 @@
+using Fleans.Domain.Connections;
 using Fleans.Domain.Exceptions;
 
 namespace Fleans.Domain;
