+++ src/Fleans/Fleans.Domain/Connections/IWorkflowConnection.cs
